# contracts/finbot-control-plane.openapi.yaml
openapi: 3.1.0
info:
  title: FinBot Administrator Control Plane API
  version: 2.0.0
  description: Authenticated Java control-plane contract for research, workflows, diagnostics, and paper-trading audit.
servers:
  - url: /api/v2
security:
  - adminSession: []
paths:
...
  /workflow-definitions/{definitionId}/versions:
    get:
      operationId: listWorkflowVersions
      parameters:
        - name: definitionId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          $ref: '#/components/responses/JsonArray'
  /workflow-definitions/{definitionId}/rollback/{targetVersionId}:
    post:
      operationId: rollbackWorkflow
      parameters:
        - name: definitionId
          in: path
          required: true
          schema: { type: string }
        - name: targetVersionId
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          $ref: '#/components/responses/JsonObject'
...
  /network/diagnostics:
    get:
      operationId: listNetworkDiagnostics
      responses:
        '200':
          $ref: '#/components/responses/JsonArray'
    post:
      operationId: startNetworkDiagnostics
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                routes:
                  type: array
                  uniqueItems: true
                  items:
                    type: string
                    enum: [FIRECRAWL, EXCHANGE_GATE, EXCHANGE_BYBIT, PUBLIC_DATA]
      responses:
        '202':
          $ref: '#/components/responses/JsonArray'
...
components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: FINBOT_SESSION
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 200 }
    CsrfToken:
      name: X-XSRF-TOKEN
      in: header
      required: true
      schema: { type: string, minLength: 1 }
...
  responses:
    JsonObject:
      description: Successful structured response.
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
    JsonArray:
      description: Successful structured collection.
      content:
        application/json:
          schema:
            type: array
            items: { type: object, additionalProperties: true }
...